Removing: Unecessary files.

+++ include/mylib/array.h
@@ -36,11 +36,11 @@
 	 * @name Operations on Array
 	 */
 	/**@{*/
-	extern array_tt array_create(int);
+	extern array_tt array_create(unsigned long int);
 	extern void array_destroy(array_tt);
-	extern int array_size(const_array_tt);
-	extern void array_set(array_tt, int, void *);
-	extern void *array_get(const_array_tt, int);
+	extern unsigned long int array_size(const_array_tt);
+	extern void array_set(array_tt, unsigned long int, void *);
+	extern void *array_get(const_array_tt, unsigned long int);
 	extern void array_shuffle(array_tt);
 	/**@}*/
